Rename to secrets_gradle_plugin
