+++ README.md
@@ -18,14 +18,14 @@ In your app-level `build.gradle` file:
 Groovy:
 ```groovy
 plugins {
-    id 'com.google.secrets_plugin' version '0.1'
+    id 'com.google.secrets_gradle_plugin' version '0.1'
 }
 ```
 
 Kotlin:
 ```groovy
 plugins {
-    id("com.google.secrets_plugin") version "0.1"
+    id("com.google.secrets_gradle_plugin") version "0.1"
 }
 ```
