+++ plugin/src/main/java/com/google/secrets_gradle_plugin/Extensions.kt
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.secrets_plugin
+package com.google.secrets_gradle_plugin
 
 import com.android.build.gradle.AppExtension
 import com.android.build.gradle.api.ApplicationVariant
