+++ sample-app/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
             android:name="com.google.android.geo.API_KEY"
             android:value="${gmpApiKey}" />
         <activity
-            android:name=".MainActivity"
+            android:name="com.google.secrets_gradle_plugin.sample.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/Theme.MyApplication.NoActionBar">
             <intent-filter>
